Reject private receivers
Starting with KiwiSDR v1.450, receivers not registering onto the public
kiwisdr.com listing, don't report anymore how many slots are open to
the public without a password, leaving no way to know if such a slot is
available; and suggesting there are, when there really aren't any.

This breaks the dispatching functionality in the context of public
use, so there is little alternative to incompatibility with private
receivers.